+++ App.tsx
@@ -1,21 +1,2 @@
-import { StatusBar } from "expo-status-bar";
-import { StyleSheet, Text, View } from "react-native";
-import React from "react";
-
-export default function App() {
-  return (
-    <View style={styles.container}>
-      <Text>TopBody</Text>
-      <StatusBar style="auto" />
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
+import App from "app/ui";
+export default App;
