+++ package.json
@@ -18,6 +18,7 @@
     "@babel/core": "^7.20.0",
     "@lad-tech/eslint-config": "^0.0.14",
     "@types/react": "~18.2.14",
+    "babel-plugin-module-resolver": "^5.0.0",
     "eslint": "^8.56.0",
     "jest": "^29.7.0",
     "prettier": "^3.2.4",
